% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle0.R
\name{angle0}
\alias{angle0}
\title{Angle Between Two 3D Vectors}
\usage{
angle0(a, b)
}
\arguments{
\item{a}{Numeric vector of dimension 3.}

\item{b}{Numeric vector of dimension 3.}
}
\value{
A numeric value corresponding to the angle between the vectors, in radians.
}
\description{
Computes the angle between two three-dimensional vectors using the dot product.
The result is restricted to the interval \eqn{[0, \pi]}.
}
\details{
The function applies a numerical correction to ensure that the argument of
\code{acos()} lies within the interval \eqn{[-1, 1]}.
}
\examples{
a <- c(1, 0, 0)
b <- c(0, 1, 0)
angle0(a, b)

}
