% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle1.R
\name{angle1}
\alias{angle1}
\title{Angle Between Two Normalized Vectors}
\usage{
angle1(a, b)
}
\arguments{
\item{a}{Numeric vector of dimension 3, normalized to unit length.}

\item{b}{Numeric vector of dimension 3, normalized to unit length.}
}
\value{
A numeric value corresponding to the angle between the vectors, in radians.
}
\description{
Computes the angle between two three-dimensional vectors that have been
previously normalized, using the dot product directly.
}
\examples{
a <- c(1, 0, 0)
b <- c(0, 1, 0)
angle1(a, b)

}
