% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglen.R
\name{anglen}
\alias{anglen}
\title{Angle Between Normalized Vectors and a Reference Vector}
\usage{
anglen(v, a, deg = TRUE)
}
\arguments{
\item{v}{Numeric matrix of dimension \code{d x n} containing normalized vectors.}

\item{a}{Numeric normalized vector of length \code{d}.}

\item{deg}{Logical. If \code{TRUE}, angles are returned in degrees; otherwise in radians.}
}
\value{
A numeric vector of length \code{n} containing the angles
between each column of \code{v} and \code{a}.
}
\description{
Computes the angle between each column of a matrix of normalized
vectors and a normalized reference vector.
}
\details{
Both \code{v} and \code{a} must contain unit vectors (i.e., Euclidean norm equal to 1).
No internal normalization is performed.
}
\examples{
v <- matrix(c(1, 0, 0,
              0, 1, 0), nrow = 3)
a <- c(1, 0, 0)
anglen(v, a, deg = TRUE)

}
