% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cad.R
\docType{data}
\name{cad}
\alias{cad}
\title{CAD model of a steel panel}
\format{
\code{mesh3d} object with the following components:
\describe{
\item{vb}{Numeric 4 x 210 matrix containing vertex coordinates in homogeneous form.}
\item{it}{Integer 3 x 256 matrix containing triangle indices.}
}
}
\source{
Theoretical CAD model of a steel panel.
}
\usage{
cad
}
\description{
Portion of the theoretical CAD model of a steel panel, represented as a \code{mesh3d}
object. The model provides the reference geometry of the panel and can be used for
geometric comparison with estimated meshes or for point cloud simulation.
}
\details{
The mesh contains 210 vertices and 256 triangular faces.
}
\examples{
data("cad", package = "dimControl")

# Show the structure
str(cad)

# Represent the CAD model
open3d() # Alternatively, use `legendplot::new3d()` to clear the current device or open a new one
shade3d(cad, col = "gray")
decorate3d()

}
\keyword{datasets}
