% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanMesh3d_rgl.R
\name{cleanMesh3d_rgl}
\alias{cleanMesh3d_rgl}
\title{Limpia una malla 3D eliminando vértices no usados y triángulos degenerados}
\usage{
cleanMesh3d_rgl(mesh, onlyFinite = TRUE, allUsed = TRUE)
}
\arguments{
\item{mesh}{Objeto de clase \code{mesh3d} que representa la malla a limpiar.}

\item{onlyFinite}{Lógico; si TRUE elimina vértices con coordenadas no finitas.}

\item{allUsed}{Lógico; si TRUE elimina vértices que no están referenciados por ninguna cara.}
}
\value{
Un objeto \code{mesh3d} limpio y con índices actualizados.
}
\description{
Esta función elimina vértices duplicados, triángulos degenerados y vértices que no
están referenciados por ninguna cara en un objeto \code{mesh3d}.
}
\details{
Reproduce la función \code{cleanMesh3d()} del paquete \strong{rgl}, desarrollada por Duncan
Murdoch, para limpiar mallas internamente sin depender de funciones no exportadas.
}
\examples{
library(rgl)

# Crear un cubo
cubo <- cube3d()

# Añadir vértice no referenciado
cubo$vb <- cbind(cubo$vb, c(10,10,10,1))

# Limpiar el cubo
cubo_clean <- cleanMesh3d_rgl(cubo)

# Comparar dimensiones
dim(cubo$vb)
dim(cubo_clean$vb)

}
\references{
Murdoch, D. (2024). \emph{rgl: 3D Visualization Using OpenGL}.
R package version correspondiente.
URL: \url{https://CRAN.R-project.org/package=rgl}
}
\author{
Duncan Murdoch (implementación original en rgl).
}
