% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoundarySegments3.R
\name{getBoundarySegments3}
\alias{getBoundarySegments3}
\title{Extrae el contorno de una malla 3D}
\usage{
getBoundarySegments3(mesh, malla = FALSE, simplify = TRUE)
}
\arguments{
\item{mesh}{Objeto de clase \code{mesh3d} que representa la malla 3D.}

\item{malla}{Valor lógico que indica si se debe devolver un objeto \code{mesh3d}
con los segmentos frontera (\code{TRUE}) o solo los índices de los vértices frontera
(\code{FALSE}). Por defecto es \code{FALSE}.}

\item{simplify}{Valor lógico que indica si se debe simplificar la malla resultante
cuando \code{malla = TRUE}. Por defecto es \code{TRUE}.}
}
\value{
Si \code{malla = FALSE}, devuelve una matriz con los índices de los vértices que
forman los bordes frontera. Si \code{malla = TRUE}, devuelve un objeto de clase
\code{mesh3d} que contiene los segmentos frontera.
}
\description{
Construye una malla de segmentos de línea correspondientes a los bordes no compartidos
(es decir, los bordes frontera) de los triángulos o cuadriláteros que componen la
malla original.
}
\details{
La función puede devolver los índices de los bordes frontera o, de forma alternativa,
generar un objeto de clase \code{mesh3d} que los contenga.
}
\examples{
# Obtener el borde de un cubo y representarlo
library(rgl)
library(data.table)

# Crear un cubo y eliminar dos caras
x <- cube3d(col = "lightblue")
x$ib <- x$ib[, -(1:2)]

# Calcular el borde
b <- getBoundarySegments3(x, malla = TRUE)

# Representar la malla y su borde
open3d()
shade3d(x, alpha = 0.2)

}
