% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getView3d.R
\name{getView3d}
\alias{getView3d}
\title{Get the Current 3D View}
\usage{
getView3d()
}
\value{
A list containing:
\itemize{
\item \code{zoom}: current zoom factor.
\item \code{userMatrix}: current user transformation matrix.
\item \code{userProjection}: current projection matrix.
}
}
\description{
Retrieves the current 3D view parameters from the active \code{rgl} device.
}
\examples{
\dontrun{
rgl::open3d()

view <- getView3d()
view
}

}
\seealso{
\code{\link[=setView3d]{setView3d()}}, \code{\link[rgl:par3d]{rgl::par3d()}}
}
