% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\docType{data}
\name{points}
\alias{points}
\title{Trozo de nube de puntos de un panel de acero}
\format{
A matrix with 14738135 rows and 3 columns:
\describe{
\item{X}{Coordenada X de cada punto}
\item{Y}{Coordenada Y de cada punto}
\item{Z}{Coordenada Z de cada punto}
}
}
\source{
Datos de escaneo 3D de un panel de acero
}
\description{
Este conjunto de datos contiene un fragmento de la nube de puntos obtenida mediante
escaneo 3D de un panel de acero. Cada fila representa un punto en el espacio con
sus coordenadas X, Y y Z.
}
