% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\docType{data}
\name{points_slave}
\alias{points_slave}
\title{Scan Slave: Trozo de nube de puntos del panel de acero}
\format{
A matrix with 7782569 rows and 3 columns:
\describe{
\item{X}{Coordenada X de cada punto}
\item{Y}{Coordenada Y de cada punto}
\item{Z}{Coordenada Z de cada punto}
}
}
\source{
Datos de escaneo 3D de un panel de acero
}
\description{
Este conjunto de datos contiene un fragmento de la nube de puntos obtenida
mediante el escaneo 3D del panel de acero con el escáner llamado \strong{Slave}.
Cada fila representa un punto en el espacio con sus coordenadas X, Y y Z.
}
