% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2deg.R
\name{rad2deg}
\alias{rad2deg}
\title{Conversion from Radians to Degrees}
\usage{
rad2deg(x)
}
\arguments{
\item{x}{A numeric value or numeric vector in radians.}
}
\value{
A numeric value or numeric vector with the angles converted to degrees.
}
\description{
Converts angular values expressed in radians to degrees.
}
\examples{
rad2deg(pi)
rad2deg(c(0, pi/2, pi))

}
