% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radDeg.R
\name{radDeg}
\alias{radDeg}
\title{Conversion from Radians to Degrees}
\usage{
radDeg(x)
}
\arguments{
\item{x}{A numeric value or numeric vector in radians.}
}
\value{
A numeric value or numeric vector with the angles converted to degrees.
}
\description{
Converts angular values expressed in radians to degrees.
}
\details{
The conversion from radians to degrees is defined as:

\deqn{
\mathrm{degrees} = \mathrm{radians} \frac{180}{\pi}
}
}
\examples{
\dontrun{
radDeg(pi)
radDeg(c(0, pi/2, pi))
}

}
