% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setView3d.R
\name{setView3d}
\alias{setView3d}
\title{Set a 3D View}
\usage{
setView3d(view)
}
\arguments{
\item{view}{A list containing 3D view parameters, typically obtained from \code{\link[=getView3d]{getView3d()}}.}
}
\value{
Invisibly returns the result of \code{\link[rgl:par3d]{rgl::par3d()}}.
}
\description{
Applies previously saved 3D view parameters to the active \code{rgl} device. The view
can be obtained using \code{\link[=getView3d]{getView3d()}}.
}
\examples{
\dontrun{
mesh <- rgl::icosahedron3d()
rgl::shade3d(mesh, color = "lightgray")

# Save the current view
view <- getView3d()

# Restore the saved view
setView3d(view)
}

}
\seealso{
\code{\link[=getView3d]{getView3d()}}, \code{\link[rgl:par3d]{rgl::par3d()}}
}
