% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{TruncatedNormal}
\alias{TruncatedNormal}
\alias{dtnorm}
\alias{qtnorm}
\alias{rtnorm}
\title{Truncated normal distribution}
\usage{
dtnorm(x, mean = 0, sd = 1, a = -Inf, b = Inf)

qtnorm(p, mean = 0, sd = 1, a = -Inf, b = Inf)

rtnorm(n, mean = 0, sd = 1, a = -Inf, b = Inf)
}
\arguments{
\item{x}{Numeric vector of values at which the density is evaluated.}

\item{mean}{Mean of the normal distribution. Default is 0.}

\item{sd}{Standard deviation of the normal distribution. Default is 1.}

\item{a}{Lower truncation point. Default is \code{-Inf}.}

\item{b}{Upper truncation point. Default is \code{Inf}.}

\item{p}{Numeric vector of probabilities.}

\item{n}{Number of observations to generate.}
}
\value{
\code{dtnorm()} returns the density evaluated at \code{x}.

\code{qtnorm()} returns the quantiles corresponding to \code{p}.

\code{rtnorm()} returns a numeric vector of \code{n} random values.
}
\description{
Density, quantile function, and random generation for the truncated
normal distribution.
}
\details{
The truncated normal distribution corresponds to a normal distribution
with mean \code{mean} and standard deviation \code{sd}, restricted to the interval
from \code{a} to \code{b}.

The functions \code{dtnorm()}, \code{qtnorm()}, and \code{rtnorm()} provide the density,
quantile function, and random generation, respectively.
}
\examples{
# Density of a truncated normal distribution
x <- seq(-2, 2, length.out = 200)
plot(
  x,
  dtnorm(x, mean = 0, sd = 1, a = -2, b = 2),
  type = "l",
  xlab = "x",
  ylab = "Density"
)

# Generate random values
set.seed(1)
values <- rtnorm(1000, mean = 0, sd = 1, a = -2, b = 2)
hist(values, probability = TRUE)

}
