% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprod.R
\name{xProd}
\alias{xProd}
\title{3D Cross Product}
\usage{
xProd(v, w)
}
\arguments{
\item{v}{Numeric vector of length 3.}

\item{w}{Numeric vector of length 3.}
}
\value{
Numeric vector of length 3 perpendicular to \code{v} and \code{w}. Its magnitude equals the
area of the parallelogram defined by both vectors, and its direction follows the
right-hand rule.
}
\description{
Computes the cross product between two three-dimensional vectors \code{v} and \code{w} in
\eqn{\mathbb{R}^3}.
}
\details{
The cross product is defined as:

\deqn{
v \times w =
\begin{pmatrix}
v_2 w_3 - v_3 w_2 \\
v_3 w_1 - v_1 w_3 \\
v_1 w_2 - v_2 w_1
\end{pmatrix}
}
}
\examples{
\dontrun{
xProd(c(1, 0, 0), c(0, 1, 0))
}

}
