% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bbox}
\alias{bbox}
\title{Calcular los límites espaciales (bounding box) de una malla 3D}
\usage{
bbox(x)
}
\arguments{
\item{x}{Objeto de clase \code{mesh3d} que contiene los vértices de la malla en el componente \code{vb}.}
}
\value{
Una matriz de 3 filas y 2 columnas con los valores mínimos y máximos de las coordenadas
de los vértices:
\itemize{
\item Cada fila corresponde a un eje (\code{x}, \code{y}, \code{z}).
\item La primera columna (\code{min}) contiene los valores mínimos.
\item La segunda columna (\code{max}) contiene los valores máximos.
}
}
\description{
Determina las coordenadas mínimas y máximas de los vértices de una malla 3D, devolviendo
una matriz con los valores extremos en los ejes \code{X}, \code{Y} y \code{Z}.
}
\details{
Internamente, la función convierte las coordenadas homogéneas de \code{x$vb} en coordenadas
euclidianas mediante \code{rgl::asEuclidean2()} y calcula los rangos por eje con \code{apply(..., range)}.
}
\examples{
require(rgl)

# Crear una malla cúbica
cube <- rgl::cube3d()

# Calcular su bounding box
bbox(cube)

}
\seealso{
\code{rgl::asEuclidean2()}
}
