% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundingBox.R
\name{boundingBox}
\alias{boundingBox}
\title{Compute the Bounding Box of a 3D Mesh}
\usage{
boundingBox(x)
}
\arguments{
\item{x}{A \code{mesh3d} object containing the mesh vertices in the \code{vb} component.}
}
\value{
A 3 x 2 matrix with rows corresponding to the \code{X}, \code{Y}, and \code{Z} axes and columns
\code{min} and \code{max}.
}
\description{
Computes the minimum and maximum coordinates of the vertices of a 3D mesh, along
the \code{X}, \code{Y}, and \code{Z} axes.
}
\details{
Homogeneous coordinates in \code{x$vb} are converted to Euclidean coordinates using
\code{\link[rgl:matrices]{rgl::asEuclidean2()}} before computing the coordinate ranges.
}
\examples{
\dontrun{
# Create a cubic mesh
cube <- rgl::cube3d()

# Compute its bounding box
boundingBox(cube)
}

}
\seealso{
\code{\link[rgl:matrices]{rgl::asEuclidean2()}}
}
