% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cad.R
\docType{data}
\name{cad}
\alias{cad}
\title{CAD object of a steel panel}
\format{
A \code{mesh3d} object (list) with 3 main components:
\describe{
\item{vb}{matrix 4 x 267 with vertex coordinates (homogeneous)}
\item{it}{matrix 3 x 343 with triangle indices}
\item{normals}{matrix 4 x 267 with vertex normals}
}
}
\source{
Theoretical CAD model of a steel panel
}
\description{
This object corresponds to a portion of the theoretical CAD model of a steel panel,
stored as a \code{mesh3d} object.
}
