% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModes.R
\name{findModes}
\alias{findModes}
\title{Calcular modas de un conjunto de datos}
\usage{
findModes(x, bw = 2, Q1 = 0.95, plot = TRUE)
}
\arguments{
\item{x}{Vector numérico. Los datos sobre los que se calcularán las modas.}

\item{bw}{Número. Ancho de banda (\emph{bandwidth}) para la estimación de densidad.
Por defecto \code{bw = 2}.}

\item{Q1}{Número entre 0 y 1. Percentil de la densidad que se usa para filtrar
los picos más relevantes. Por defecto \code{Q1 = 0.95}.}

\item{plot}{Lógico. Si es \code{TRUE}, se genera un gráfico de la densidad con los picos
detectados. Por defecto es \code{TRUE}.}
}
\value{
Una matriz con tantas filas como modas detectadas y dos columnas: \code{min}
y \code{max}, que representan el rango en X de cada moda.
}
\description{
Esta función detecta los picos (modas) en un conjunto de datos \code{x} usando estimación
de densidad. Para cada moda, determina el rango en el eje X donde la densidad desciende
hacia los mínimos adyacentes. También permite representar la densidad y los picos
detectados.
}
\examples{
set.seed(123) # Semilla
x <- c(
  rnorm(200, mean = 2, sd = 0.3),
  rnorm(5, mean = 3.5, sd = 0.1),
  rnorm(200, mean = 5, sd = 0.3)
)

# Calcular y representar modas
modas <- findModes(x, bw = 0.2, Q1 = 0.5, plot = TRUE)

# Rangos en X de cada moda
modas

}
