% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModes.R
\name{findModes}
\alias{findModes}
\title{Detect Modes in a Dataset}
\usage{
findModes(x, bw = 2, Q1 = 0.95, plot = TRUE)
}
\arguments{
\item{x}{Numeric vector. The data for which modes will be calculated.}

\item{bw}{Numeric. Bandwidth for the density estimation. Default is \code{bw = 2}.}

\item{Q1}{Numeric between 0 and 1. Percentile of the density used to filter the
most relevant peaks. Default is \code{Q1 = 0.95}.}

\item{plot}{Logical. If \code{TRUE}, generates a plot of the density with detected modes.
Default is \code{TRUE}.}
}
\value{
A matrix with one row per detected mode and two columns (\code{min} and \code{max})
representing the X-axis range of each mode.
}
\description{
Detects peaks (modes) in a numeric dataset \code{x} using kernel density estimation.
For each mode, it determines the range along the X-axis where the density decreases
toward the adjacent minima. The function can also plot the density and the detected modes.
}
\examples{
set.seed(123) # Seed
x <- c(
  rnorm(200, mean = 2, sd = 0.3),
  rnorm(5, mean = 3.5, sd = 0.1),
  rnorm(200, mean = 5, sd = 0.3)
)

# Compute and plot modes
modes <- findModes(x, bw = 0.2, Q1 = 0.5, plot = TRUE)

# X-axis ranges for each mode
modes

}
