% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getview3d.R
\name{getview3d}
\alias{getview3d}
\title{Obtener la vista 3D actual}
\usage{
getview3d()
}
\value{
Lista con \code{zoom}, \code{userMatrix} y \code{userProjection}.
}
\description{
Recupera los parámetros actuales de la vista 3D del dispositivo activo de rgl.
}
