% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getView3d.R
\name{getView3d}
\alias{getView3d}
\title{Get the Current 3D View}
\usage{
getView3d()
}
\value{
A list containing the current \code{zoom}, \code{userMatrix}, and \code{userProjection}.
}
\description{
Retrieves the current 3D view parameters from the active \code{rgl} device.
}
