% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMesh.R
\name{sampleMesh}
\alias{sampleMesh}
\title{Sampling points over a triangular mesh}
\usage{
sampleMesh(mesh, n, shuffle = FALSE)
}
\arguments{
\item{mesh}{A list representing a triangular mesh, containing:
\itemize{
\item \code{vb}: a 3xN (or 4xN homogeneous) matrix with the mesh vertex coordinates.
\item \code{it}: a 3xM matrix with the vertex indices for each triangle (each column is a triangle).
}}

\item{n}{Integer greater than 0. Total number of points to generate.}

\item{shuffle}{Logical. Whether to shuffle the generated points randomly.}
}
\value{
A numeric matrix of dimension 3 x n, where each column represents
a point sampled in 3D space.
}
\description{
Generates random points uniformly distributed over a triangular mesh.
Each triangle receives a number of points proportional to its area, and the points
are generated using barycentric coordinates.
}
\details{
Internally, this function uses \code{sampleTriangle()} to sample points
within each triangle.

If the mesh uses homogeneous coordinates (4th row in \code{vb}), points are converted
to 3D by dividing by the w component. The points are returned as columns
of a 3xN matrix. If \code{shuffle = TRUE}, the point order is shuffled randomly.
}
\examples{
\dontrun{
# Create a triangular sphere
mesh <- Rvcg::vcgSphere(1, subdiv = 1)
rgl::wire3d(mesh) # visualize the mesh
# Sample 10000 points over the mesh
pts <- sampleMesh(mesh, 10000, shuffle = TRUE)
# Plot the sampled points
rgl::points3d(t(pts), col = "red", size = 5)
}

}
