% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setview3d.R
\name{setview3d}
\alias{setview3d}
\title{Establecer una vista 3D}
\usage{
setview3d(view)
}
\arguments{
\item{view}{Lista obtenida con \code{getview3d()}.}
}
\description{
Aplica una vista previamente guardada al dispositivo rgl.
}
