% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setView3d.R
\name{setView3d}
\alias{setView3d}
\title{Set a 3D View}
\usage{
setView3d(view)
}
\arguments{
\item{view}{A list obtained from \code{getview3d()}.}
}
\description{
Applies a previously saved 3D view to the active \code{rgl} device.
}
