% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCloud.R
\name{simulateFloor}
\alias{simulateFloor}
\title{Simulate floor point cloud}
\usage{
simulateFloor(object, n, gap, margin, rgen = rtnorm, ...)
}
\arguments{
\item{object}{Numeric matrix with three columns (\code{X}, \code{Y}, \code{Z}) containing
the object point cloud.}

\item{n}{Number of floor points to simulate.}

\item{gap}{Non-negative numeric value specifying the vertical distance
between the lowest point of the object and the reference floor height.}

\item{margin}{Positive numeric value specifying the width of the floor region
around the object in the X and Y directions.}

\item{rgen}{Random number generator used to perturb the floor height. By
default, \code{\link[=rtnorm]{rtnorm()}} is used. \code{\link[=rtri]{rtri()}} is also provided by \code{dimControl},
although any random number generator can be supplied.}

\item{...}{Additional arguments passed to \code{rgen}.}
}
\value{
A numeric matrix with \code{n} rows and three columns (\code{X}, \code{Y}, \code{Z}) containing
the simulated floor point cloud.
}
\description{
Simulates a floor point cloud around the bounding box of a 3D object point
cloud.
}
\details{
The floor is simulated in four rectangular regions surrounding the X-Y
bounding box of \code{object}: left, right, bottom, and top.

The reference floor height is defined as the minimum Z coordinate of the
object minus \code{gap}. Random noise generated by \code{rgen} is then added to this
reference height.

\code{dimControl} provides \code{\link[=rtnorm]{rtnorm()}} and \code{\link[=rtri]{rtri()}} as random number generators
that can be used to model the random perturbation. Other random number
generators can also be supplied through \code{rgen}, provided that their first
argument specifies the number of values to generate. Distribution-specific
parameters are passed through \code{...}.
}
\seealso{
\code{\link[=simulateCloud]{simulateCloud()}}, \code{\link[=rtnorm]{rtnorm()}}, \code{\link[=rtri]{rtri()}}
}
