% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceDistance.R
\name{surfaceDistance}
\alias{surfaceDistance}
\title{Calcular la distancia sobre una superficie 3D}
\usage{
surfaceDistance(segmento)
}
\arguments{
\item{segmento}{Matriz o data frame con tres columnas (\code{x}, \code{y}, \code{z}) que representan
las coordenadas de los puntos consecutivos en el espacio 3D.}
}
\value{
Un valor numérico que indica la distancia total recorrida sobre la superficie.
}
\description{
Esta función calcula la longitud total de una trayectoria tridimensional, sumando
las distancias euclidianas entre puntos consecutivos. Es útil para estimar la distancia
desarrollada sobre una superficie 3D.
}
